# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APLICACION CREADA PARA GESTION DE TALENTO HUMANO EN EL GAD MUNICIPAL DEL CANTON COLTA",
        "description": "Esta API fue realizada con el propósito de ser utilizada en los sistemas del GAD de Colta. El uso no autorizado está prohibido.",
        "termsOfService": "https://tu-url.com/terminos",
        "contact": {
            "name": "Angel Melendres & Edison Gualle",
            "url": "https://tu-url.com/contacto",
            "email": "correo@ejemplo.com"
        },
        "license": {
            "name": "Licencia de grupo gym",
            "url": "https://tu-url.com/licencia"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/capacitaciones": {
            "get": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Listar todas las capacitaciones",
                "operationId": "ac32844afab8477b45eec8780e705d4a",
                "responses": {
                    "200": {
                        "description": "Lista de todas las capacitaciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Crear una nueva capacitación",
                "operationId": "d54f91fbe4c6ee19cf5f02989d55b44f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Nombre de la capacitación"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción de la capacitación"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Capacitación creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/capacitaciones/{id}": {
            "get": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Obtener detalles de una capacitación por ID",
                "operationId": "eb58123a5bba7361799e84fbe2f1884a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la capacitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la capacitación"
                    },
                    "404": {
                        "description": "Capacitación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Actualizar una capacitación por ID",
                "operationId": "8ec9827c6b99ceca74cb178a51112a6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la capacitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Nuevo nombre de la capacitación"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Nueva descripción de la capacitación"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Capacitación actualizada exitosamente"
                    },
                    "404": {
                        "description": "Capacitación no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Eliminar una capacitación por ID",
                "operationId": "f228fe6fe65d6b23e9f42a70aec4f569",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la capacitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Capacitación eliminada exitosamente"
                    },
                    "404": {
                        "description": "Capacitación no encontrada"
                    }
                }
            }
        },
        "/capacitaciones/nombre/{nombre}": {
            "get": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Listar capacitaciones por nombre",
                "operationId": "1a19769ae31053662b3610fea38c3fa0",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "description": "Nombre de la capacitación",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de capacitaciones por nombre"
                    }
                }
            }
        },
        "/capacitaciones/fecha/{fecha}": {
            "get": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Listar capacitaciones por fecha",
                "operationId": "c99081a74f30a5e792e917af322583b3",
                "parameters": [
                    {
                        "name": "fecha",
                        "in": "path",
                        "description": "Fecha de la capacitación",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de capacitaciones por fecha"
                    }
                }
            }
        },
        "/capacitaciones/rango-fechas/{fechaInicio}/{fechaFin}": {
            "get": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Listar capacitaciones por rango de fechas",
                "operationId": "c107bc31812e88d7ddace788c9aa72e8",
                "parameters": [
                    {
                        "name": "fechaInicio",
                        "in": "path",
                        "description": "Fecha de inicio del rango",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fechaFin",
                        "in": "path",
                        "description": "Fecha de fin del rango",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de capacitaciones por rango de fechas"
                    }
                }
            }
        },
        "/capacitaciones/asignacion-empleado-capacitacion": {
            "post": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Asignar un empleado a una capacitación",
                "operationId": "cdf0c1cf4fa35edf1029c8830f43f457",
                "parameters": [
                    {
                        "name": "idEmpleado",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idCapacitacion",
                        "in": "path",
                        "description": "ID de la capacitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asignación creada exitosamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Empleado o capacitación no encontrados"
                    }
                }
            }
        },
        "/capacitaciones/asignacion-empleado-capacitacion/{idEmpleado}/{idCapacitacion}": {
            "put": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Actualizar asignación de un empleado a una capacitación",
                "operationId": "1369012ded0c2d6bc63971344560fc07",
                "parameters": [
                    {
                        "name": "idEmpleado",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idCapacitacion",
                        "in": "path",
                        "description": "ID de la capacitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Asignación actualizada exitosamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Asignación de capacitación no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Eliminar asignación de un empleado a una capacitación",
                "operationId": "24577d58539484bd2da67e482560ca0f",
                "parameters": [
                    {
                        "name": "idEmpleado",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idCapacitacion",
                        "in": "path",
                        "description": "ID de la capacitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asignación de capacitación eliminada exitosamente"
                    },
                    "404": {
                        "description": "Asignación de capacitación no encontrada"
                    }
                }
            }
        },
        "/capacitaciones/capacitaciones-por-empleado/id/{idEmpleado}": {
            "get": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Listar capacitaciones realizadas por un empleado",
                "operationId": "aec9273ddbc9268e642e4943c19432b4",
                "parameters": [
                    {
                        "name": "idEmpleado",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de capacitaciones realizadas por el empleado"
                    },
                    "404": {
                        "description": "Empleado no encontrado"
                    }
                }
            }
        },
        "/capacitaciones/empleados-por-capacitacion/id/{idCapacitacion}": {
            "get": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Listar empleados participantes en una capacitación",
                "operationId": "874926b68d23a487a7c4ae17cc4287e4",
                "parameters": [
                    {
                        "name": "idCapacitacion",
                        "in": "path",
                        "description": "ID de la capacitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de empleados participantes en la capacitación"
                    },
                    "404": {
                        "description": "Capacitación no encontrada"
                    }
                }
            }
        },
        "/capacitaciones/capacitaciones-no-realizadas-por-empleado/id/{idEmpleado}": {
            "get": {
                "tags": [
                    "Capacitaciones"
                ],
                "summary": "Listar capacitaciones no realizadas por un empleado",
                "operationId": "9aa4b1b20e425f22900d0690c25ee134",
                "parameters": [
                    {
                        "name": "idEmpleado",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de capacitaciones no realizadas por el empleado"
                    },
                    "404": {
                        "description": "Empleado no encontrado"
                    }
                }
            }
        }
    }
}